Refactor subscribing to appBridge
