+++ apps/saleor-app-checkout/frontend/components/elements/AppProvider/ClientAppBridgeProvider.tsx
@@ -1,10 +1,10 @@
 import { createGraphqlClient } from "@/saleor-app-checkout/frontend/misc/client";
 import createSafeContext from "@/saleor-app-checkout/frontend/misc/createSafeContext";
-import { actions, AppBridge } from "@saleor/app-sdk/app-bridge";
-import { useRouter } from "next/router";
+import { AppBridge } from "@saleor/app-sdk/app-bridge";
 import { Provider as UrqlProvider } from "urql";
-import { ReactNode, useMemo, useState, useEffect } from "react";
-import { getRawAppPath } from "@/saleor-app-checkout/frontend/utils";
+import { ReactNode, useMemo } from "react";
+import { useSynchronizedAppBridgePaths } from "./useSynchronizedAppBridgePaths";
+import { useSubscribeToIsAuthorized } from "./useSubscribeToIsAuthorized";
 
 interface IAppContext {
   app: AppBridge;
@@ -16,9 +16,8 @@ export { useAppContext };
 
 export const ClientAppBridgeProvider = ({ children }: { children: ReactNode }) => {
   const app = useMemo(() => new AppBridge(), []);
-
-  const router = useRouter();
-  const [isAuthorized, setIsAuthorized] = useState(!!app.getState()?.token);
+  const isAuthorized = useSubscribeToIsAuthorized(app);
+  useSynchronizedAppBridgePaths(app);
 
   // @todo use `saleorApiUrl`
   const domain = app.getState().domain;
@@ -27,43 +26,6 @@ export const ClientAppBridgeProvider = ({ children }: { children: ReactNode }) =
     () => createGraphqlClient(saleorApiUrl, app.getState().token),
     [app, saleorApiUrl]
   );
-
-  useEffect(() => {
-    if (app) {
-      setIsAuthorized(!!app.getState().token);
-
-      const unsubscribe = app.subscribe("handshake", (payload) => {
-        setIsAuthorized(!!payload.token);
-      });
-
-      return () => {
-        unsubscribe();
-      };
-    }
-  }, [app]);
-
-  useEffect(() => {
-    const handleRouteChange = (path: string) => {
-      const rawPath = getRawAppPath(path);
-      void app.dispatch(
-        actions.Redirect({
-          to: `/apps/${encodeURIComponent(app.getState().id)}/app${rawPath}`,
-        })
-      );
-    };
-
-    const unsubscribe = app.subscribe("redirect", ({ path }) => {
-      if (path && path !== router.pathname) {
-        void router.push(path);
-      }
-    });
-    router.events.on("routeChangeComplete", handleRouteChange);
-
-    return () => {
-      unsubscribe();
-      router.events.off("routeChangeComplete", handleRouteChange);
-    };
-  }, [app, router]);
 
   const appContext = useMemo(() => ({ app, isAuthorized }), [app, isAuthorized]);
